Reorganize `Menu` import
